# docs/protocol/http/openapi-v1.json
{
  "components": {
    "parameters": {
...
      "RequestId": {
        "in": "header",
        "name": "X-Request-ID",
        "required": true,
        "schema": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "responses": {
      "Problem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        },
        "description": "Stable typed failure.",
        "headers": {
          "X-Request-ID": {
            "description": "Matches the request and JSON body request_id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        }
      }
    },
    "schemas": {
      "AgentApiConfigListResponse": {
        "additionalProperties": false,
        "description": "Return all non-secret Provider configurations.",
        "properties": {
          "configs": {
            "items": {
              "$ref": "#/components/schemas/ModelApiConfig"
            },
            "title": "Configs",
            "type": "array"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "configs"
        ],
        "title": "AgentApiConfigListResponse",
        "type": "object"
      },
      "AgentApiConfigResponse": {
        "additionalProperties": false,
        "description": "Return one non-secret Provider configuration.",
        "properties": {
          "config": {
            "$ref": "#/components/schemas/ModelApiConfig"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "config"
        ],
        "title": "AgentApiConfigResponse",
        "type": "object"
      },
      "AgentRunStatus": {
        "description": "Represent the only persisted lifecycle states for an Agent run.",
        "enum": [
          "RUNNING",
          "COMPLETED",
          "FAILED",
          "LIMIT_REACHED",
          "CANCELLED"
        ],
        "title": "AgentRunStatus",
        "type": "string"
      },
      "AgentTurnRequest": {
        "additionalProperties": false,
        "description": "Validate the only secret Protocol payload accepted for an Agent turn.",
        "properties": {
          "api_config_id": {
            "format": "uuid",
            "title": "Api Config Id",
            "type": "string"
          },
          "api_key_b64": {
            "title": "Api Key B64",
            "type": "string"
          },
          "api_key_credential_id": {
            "format": "uuid",
            "title": "Api Key Credential Id",
            "type": "string"
          },
          "conversation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Conversation Id"
          },
          "ssh_session_id": {
            "format": "uuid",
            "title": "Ssh Session Id",
            "type": "string"
          },
          "user_message": {
            "maxLength": 65536,
            "minLength": 1,
            "title": "User Message",
            "type": "string"
          }
        },
        "required": [
          "ssh_session_id",
          "api_config_id",
          "api_key_credential_id",
          "api_key_b64",
          "user_message"
        ],
        "title": "AgentTurnRequest",
        "type": "object"
      },
      "AgentTurnResponse": {
        "additionalProperties": false,
        "description": "Return the complete non-streaming Agent result with HTTP correlation.",
        "properties": {
          "agent_run_id": {
            "description": "Run which produced this result.",
            "format": "uuid",
            "title": "Agent Run Id",
            "type": "string"
          },
          "conversation_id": {
            "description": "Conversation receiving the turn.",
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Stable terminal failure code, if any.",
            "title": "Error Code"
          },
          "final_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Final model text when successfully available.",
            "title": "Final Text"
          },
          "react_iteration": {
            "description": "Completed Tool Call to ToolMessage loops.",
            "maximum": 128.0,
            "minimum": 0.0,
            "title": "React Iteration",
            "type": "integer"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AgentRunStatus",
            "description": "Terminal run status."
          }
        },
        "required": [
          "conversation_id",
          "agent_run_id",
          "status",
          "final_text",
          "react_iteration",
          "error_code",
          "request_id"
        ],
        "title": "AgentTurnResponse",
        "type": "object"
      },
      "ApiType": {
        "description": "Select the one explicitly configured OpenAI-compatible API surface.",
        "enum": [
          "CHAT_COMPLETIONS",
          "RESPONSES"
        ],
        "title": "ApiType",
        "type": "string"
      },
...
      "DeleteResponse": {
        "additionalProperties": false,
        "description": "Return the determined deletion outcome for an identity.",
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "deleted"
        ],
        "title": "DeleteResponse",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "ModelApiConfig": {
        "additionalProperties": false,
        "description": "Represent one persisted API configuration without secret material.",
        "properties": {
          "api_config_id": {
            "description": "Stable configuration identity.",
            "format": "uuid",
            "title": "Api Config Id",
            "type": "string"
          },
          "api_key_secret_ref": {
            "description": "Opaque Rust Vault reference; never the API key plaintext.",
            "format": "uuid",
            "title": "Api Key Secret Ref",
            "type": "string"
          },
          "api_type": {
            "$ref": "#/components/schemas/ApiType",
            "description": "Explicit API selection; automatic probing is forbidden."
          },
          "base_url": {
            "description": "Normalized HTTP(S) base URL supplied to ChatOpenAI.",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Base Url",
            "type": "string"
          },
          "created_at": {
            "description": "UTC creation timestamp.",
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "description": "User-facing label for this provider configuration.",
            "maxLength": 80,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "description": "Whether new Agent runs may use this configuration.",
            "title": "Enabled",
            "type": "boolean"
          },
          "model": {
            "description": "Provider model identifier passed without inference.",
            "maxLength": 255,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "updated_at": {
            "description": "UTC last-update timestamp.",
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "display_name",
          "api_type",
          "base_url",
          "model",
          "api_key_secret_ref",
          "api_config_id",
          "created_at",
          "updated_at"
        ],
        "title": "ModelApiConfig",
        "type": "object"
      },
      "ModelApiConfigInput": {
        "additionalProperties": false,
        "description": "Validate non-secret model configuration supplied by a trusted Core caller.",
        "properties": {
          "api_key_secret_ref": {
            "description": "Opaque Rust Vault reference; never the API key plaintext.",
            "format": "uuid",
            "title": "Api Key Secret Ref",
            "type": "string"
          },
          "api_type": {
            "$ref": "#/components/schemas/ApiType",
            "description": "Explicit API selection; automatic probing is forbidden."
          },
          "base_url": {
            "description": "Normalized HTTP(S) base URL supplied to ChatOpenAI.",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Base Url",
            "type": "string"
          },
          "display_name": {
            "description": "User-facing label for this provider configuration.",
            "maxLength": 80,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "description": "Whether new Agent runs may use this configuration.",
            "title": "Enabled",
            "type": "boolean"
          },
          "model": {
            "description": "Provider model identifier passed without inference.",
            "maxLength": 255,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          }
        },
        "required": [
          "display_name",
          "api_type",
          "base_url",
          "model",
          "api_key_secret_ref"
        ],
        "title": "ModelApiConfigInput",
        "type": "object"
      },
...
      "ProblemDetails": {
        "additionalProperties": false,
        "description": "Expose one bounded machine-readable failure without raw internals.",
        "properties": {
          "details": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Details",
            "type": "object"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "error_code",
          "message",
          "request_id",
          "details"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Loopback-only API owned by the Rust-managed packaged child. The Runtime WebSocket is specified separately.",
    "title": "Harness Shell Private Python Runtime API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/agent/api-configs": {
      "get": {
        "description": "Return all non-secret Provider configurations.",
        "operationId": "list_api_configs_v1_agent_api_configs_get",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentApiConfigListResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "List Api Configs"
      },
      "post": {
        "description": "Persist one Provider configuration containing only a Vault reference.",
        "operationId": "create_api_config_v1_agent_api_configs_post",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelApiConfigInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentApiConfigResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Create Api Config"
      }
    },
    "/v1/agent/api-configs/{api_config_id}": {
      "delete": {
        "description": "Delete Sidecar metadata without changing Rust Vault ownership.",
        "operationId": "delete_api_config_v1_agent_api_configs__api_config_id__delete",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "in": "path",
            "name": "api_config_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Api Config Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Delete Api Config"
      },
      "get": {
        "description": "Return one non-secret Provider configuration.",
        "operationId": "get_api_config_v1_agent_api_configs__api_config_id__get",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "in": "path",
            "name": "api_config_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Api Config Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentApiConfigResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Get Api Config"
      },
      "patch": {
        "description": "Replace one complete non-secret Provider configuration.",
        "operationId": "update_api_config_v1_agent_api_configs__api_config_id__patch",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "in": "path",
            "name": "api_config_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Api Config Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelApiConfigInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentApiConfigResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Update Api Config"
      }
    },
    "/v1/agent/turns": {
      "post": {
        "description": "Run one complete non-streaming Agent turn with transient API key bytes.",
        "operationId": "run_agent_turn_v1_agent_turns_post",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentTurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTurnResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Run Agent Turn"
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:{port}",
      "variables": {
        "port": {
          "default": "8765"
        }
      }
    }
  ]
}